% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment}
\alias{augment}
\title{Augment data according to a tidied meta-analysis model.}
\usage{
augment(x)
}
\arguments{
\item{x}{A meta-analysis model produced by metafor's rma function.}
}
\description{
Given a metafor model, add columns to the original dataset such as
predictions, residuals and cluster assignments.
}
\details{
As far as I can see, there is no way to call the original dataframe from the
}
