% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ma.R
\docType{data}
\name{example_ma}
\alias{example_ma}
\title{Example meta-analysis}
\format{An object of class \code{rma.uni} (inherits from \code{rma}) of length 72.}
\usage{
example_ma
}
\description{
An example \code{rma} object, using a dataset from the \code{metafor::} package.
}
\details{
See data-raw/test_ma.R and \code{[metafor::dat.normand1999]} for more information.
}
\keyword{datasets}
