% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance}
\alias{glance}
\title{Construct a single-row summary "glance" of a meta-analysis model.}
\usage{
glance(x)
}
\arguments{
\item{x}{A meta-analysis model produced by metafor's rma function.}
}
\description{
This function is analogous to the broom::tidy function but for meta-analysis
models.
glance methods always return either a one-row data frame, or NULL.
}
