% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{Tidy the result of a meta-analysis into a summary data.frame.}
\usage{
tidy(rma)
}
\arguments{
\item{x}{A meta-analysis model produced by metafor's rma function.}
}
\description{
This function is analogous to the broom::tidy function but for meta-analysis
models.
}
\details{
From the broom tidy documentation:
The output of tidy is always a data.frame with disposable row names.
It is therefore suited for further manipulation by packages like dplyr,
reshape2, ggplot2 and ggvis.

This aims to produce a data frame that summarises the model's statistical
findings.
}
